chore(core-manager): rearrange folder structure
