+++ __tests__/unit/core-manager/worker/generate-log-zip.test.ts
@@ -1,11 +1,10 @@
 import "jest-extended";
 
 import { Database } from "@packages/core-manager/src/database/database";
-import { GenerateLogGz } from "@packages/core-manager/src/workers/actions/generate-log-gz";
-import { createReadStream } from "fs-extra";
+import { GenerateLogZip } from "@packages/core-manager/src/workers/generate-log-zip";
+import { existsSync } from "fs-extra";
 import { join } from "path";
 import { dirSync, setGracefulCleanup } from "tmp";
-import zlib from "zlib";
 
 jest.mock("@packages/core-manager/src/database/database");
 
@@ -31,10 +30,10 @@ describe("Generate Log", () => {
         const spyOnGetAllIterator = jest.spyOn(Database.prototype, "getAllIterator").mockReturnValue(mockIterator);
 
         // @ts-ignore
-        const generateLog = new GenerateLogGz({
+        const generateLog = new GenerateLogZip({
             databaseFilePath: "path/to/db",
             schema: { tables: [] },
-            logFileName: "test.log.gz",
+            logFileName: "test.zip",
             query: {},
         });
 
@@ -44,16 +43,6 @@ describe("Generate Log", () => {
 
         expect(spyOnGetAllIterator).toHaveBeenCalled();
 
-        const stream = createReadStream(join(process.env.CORE_PATH_DATA!, "log-archive", "test.log.gz")).pipe(
-            zlib.createGunzip(),
-        );
-
-        await new Promise((resolve) => {
-            stream.on("readable", () => {
-                resolve();
-            });
-        });
-
-        expect(stream.read().toString()).toEqual("[2020-12-14 12:20:05.000] INFO : log message\n");
+        expect(existsSync(join(process.env.CORE_PATH_DATA!, "log-archive", "test.zip"))).toBeTrue();
     });
 });
