--- packages/core-manager/src/workers/actions/streams/index.ts
@@ -1,1 +1,0 @@
-export * from "./log-transform-stream";
